404 errors should not be raised as 500
This error has not been reproduced on heroku or locally, but is appearing on stage and live. This fix is based on the stacktrace (which says one of the context processors was trying to find out the view name of urls that didn't exist thereby causing 500 errors). The context processor has been removed and the functionality has been reimplemented in the three views that actually need to have that variable in the context.