CI-108: Turn off custom tags for GA temporarily.
The spec for the data to be sent in these tags has been changed.
We don't have time to update and test the changes before release, so we are temporarily turning off this feature in preference to sending incorrect and potentially confusing data.

General GA data (eg tracking of the URLs that that user visits) will not be affected by this commit and will continue to send as normal.